+++ src/main/java/org/elasticsearch/search/aggregations/reducers/derivative/DerivativeReducer.java
@@ -22,37 +22,29 @@ package org.elasticsearch.search.aggregations.reducers.derivative;
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 
-import org.elasticsearch.ElasticsearchIllegalStateException;
+
 import org.elasticsearch.common.Nullable;
-import org.elasticsearch.common.ParseField;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.search.SearchParseException;
 import org.elasticsearch.search.aggregations.Aggregation;
-import org.elasticsearch.search.aggregations.AggregationExecutionException;
 import org.elasticsearch.search.aggregations.Aggregator;
 import org.elasticsearch.search.aggregations.InternalAggregation;
 import org.elasticsearch.search.aggregations.InternalAggregation.ReduceContext;
 import org.elasticsearch.search.aggregations.InternalAggregation.Type;
 import org.elasticsearch.search.aggregations.InternalAggregations;
-import org.elasticsearch.search.aggregations.InvalidAggregationPathException;
 import org.elasticsearch.search.aggregations.bucket.histogram.InternalHistogram;
-import org.elasticsearch.search.aggregations.metrics.InternalNumericMetricsAggregation;
-import org.elasticsearch.search.aggregations.reducers.InternalSimpleValue;
-import org.elasticsearch.search.aggregations.reducers.Reducer;
-import org.elasticsearch.search.aggregations.reducers.ReducerFactory;
-import org.elasticsearch.search.aggregations.reducers.ReducerStreams;
+import org.elasticsearch.search.aggregations.reducers.*;
 import org.elasticsearch.search.aggregations.support.AggregationContext;
-import org.elasticsearch.search.aggregations.support.AggregationPath;
 import org.elasticsearch.search.aggregations.support.format.ValueFormatter;
 import org.elasticsearch.search.aggregations.support.format.ValueFormatterStreams;
-import org.elasticsearch.search.internal.SearchContext;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+
+import static org.elasticsearch.search.aggregations.reducers.BucketHelpers.GapPolicy;
+import static org.elasticsearch.search.aggregations.reducers.BucketHelpers.resolveBucketValue;
 
 public class DerivativeReducer extends Reducer {
 
@@ -105,7 +97,7 @@ public class DerivativeReducer extends Reducer {
         List newBuckets = new ArrayList<>();
         Double lastBucketValue = null;
         for (InternalHistogram.Bucket bucket : buckets) {
-            Double thisBucketValue = resolveBucketValue(histo, bucket);
+            Double thisBucketValue = resolveBucketValue(histo, bucket, bucketsPaths()[0], gapPolicy);
             if (lastBucketValue != null) {
                 double diff = thisBucketValue - lastBucketValue;
 
@@ -120,40 +112,6 @@ public class DerivativeReducer extends Reducer {
             lastBucketValue = thisBucketValue;
         }
         return factory.create(histo.getName(), newBuckets, histo);
-    }
-
-    private Double resolveBucketValue(InternalHistogram<? extends InternalHistogram.Bucket> histo, InternalHistogram.Bucket bucket) {
-        try {
-            Object propertyValue = bucket.getProperty(histo.getName(), AggregationPath.parse(bucketsPaths()[0])
-                    .getPathElementsAsStringList());
-            if (propertyValue == null) {
-                throw new AggregationExecutionException(DerivativeParser.BUCKETS_PATH.getPreferredName()
-                        + " must reference either a number value or a single value numeric metric aggregation");
-            } else {
-                double value;
-                if (propertyValue instanceof Number) {
-                    value = ((Number) propertyValue).doubleValue();
-                } else if (propertyValue instanceof InternalNumericMetricsAggregation.SingleValue) {
-                    value = ((InternalNumericMetricsAggregation.SingleValue) propertyValue).value();
-                } else {
-                    throw new AggregationExecutionException(DerivativeParser.BUCKETS_PATH.getPreferredName()
-                            + " must reference either a number value or a single value numeric metric aggregation");
-                }
-                if (Double.isInfinite(value) || Double.isNaN(value)) {
-                    switch (gapPolicy) {
-                    case INSERT_ZEROS:
-                        return 0.0;
-                    case IGNORE:
-                    default:
-                        return Double.NaN;
-                    }
-                } else {
-                    return value;
-                }
-            }
-        } catch (InvalidAggregationPathException e) {
-            return null;
-        }
     }
 
     @Override
@@ -185,57 +143,5 @@ public class DerivativeReducer extends Reducer {
             return new DerivativeReducer(name, bucketsPaths, formatter, gapPolicy, metaData);
         }
 
-    }
-
-    public static enum GapPolicy {
-        INSERT_ZEROS((byte) 0, "insert_zeros"), IGNORE((byte) 1, "ignore");
-
-        public static GapPolicy parse(SearchContext context, String text) {
-            GapPolicy result = null;
-            for (GapPolicy policy : values()) {
-                if (policy.parseField.match(text)) {
-                    if (result == null) {
-                        result = policy;
-                    } else {
-                        throw new ElasticsearchIllegalStateException("Text can be parsed to 2 different gap policies: text=[" + text
-                                + "], " + "policies=" + Arrays.asList(result, policy));
-                    }
-                }
-            }
-            if (result == null) {
-                final List<String> validNames = new ArrayList<>();
-                for (GapPolicy policy : values()) {
-                    validNames.add(policy.getName());
-                }
-                throw new SearchParseException(context, "Invalid gap policy: [" + text + "], accepted values: " + validNames);
-            }
-            return result;
-        }
-
-        private final byte id;
-        private final ParseField parseField;
-
-        private GapPolicy(byte id, String name) {
-            this.id = id;
-            this.parseField = new ParseField(name);
-        }
-
-        public void writeTo(StreamOutput out) throws IOException {
-            out.writeByte(id);
-        }
-
-        public static GapPolicy readFrom(StreamInput in) throws IOException {
-            byte id = in.readByte();
-            for (GapPolicy gapPolicy : values()) {
-                if (id == gapPolicy.id) {
-                    return gapPolicy;
-                }
-            }
-            throw new IllegalStateException("Unknown GapPolicy with id [" + id + "]");
-        }
-
-        public String getName() {
-            return parseField.getPreferredName();
-        }
     }
 }
